old/allrgb/requirements.txt
astroid==1.4.9
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
olefile==0.44
Pillow==4.0.0
pylint==1.6.5
six==1.10.0
wrapt==1.10.10
